+++ src/js/init/initDisplayDate.js
@@ -1,4 +1,5 @@
 import getCurrentDate from "../utils/getCurrentDate.js";
+import monthNames from "../constants/monthNames.js";
 
 export default () => {
   const { year, month, monthName } = getCurrentDate();
